additional test cases for non-event sourced command target scheduling
