+++ Domain.Sql.Tests/SqlCommandSchedulerTests_EventSourced.cs
@@ -68,7 +68,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
                          .UseDependency<IEventSourcedRepository<Order>>(t => orderRepository)
                          .UseDependency<IEventSourcedRepository<CustomerAccount>>(t => accountRepository);
 
-            ConfigureScheduler(configuration);
+            Configure(configuration);
 
             disposables.Add(ConfigurationContext.Establish(configuration));
 
@@ -168,7 +168,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_clock_is_advanced_then_commands_are_not_triggered_that_have_not_become_due()
+        public override async Task When_a_clock_is_advanced_then_commands_are_not_triggered_that_have_not_become_due()
         {
             // arrange
             var shipmentId = Any.AlphanumericString(8, 8);
@@ -191,7 +191,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task Scheduled_commands_are_delivered_immediately_if_a_past_due_time_is_specified_on_the_realtime_clock()
+        public override async Task Scheduled_commands_are_delivered_immediately_if_past_due_per_the_domain_clock()
         {
             // arrange
             var order = CommandSchedulingTests_EventSourced.CreateOrder();
@@ -209,7 +209,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
       
         [Test]
-        public async Task Scheduled_commands_are_delivered_immediately_if_a_past_due_time_is_specified_on_a_scheduler_clock()
+        public override async Task Scheduled_commands_are_delivered_immediately_if_past_due_per_the_scheduler_clock()
         {
             // arrange
             var order = CommandSchedulingTests_EventSourced.CreateOrder();
@@ -218,7 +218,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
 
             // act
             var shipOn = clockRepository.ReadClock(clockName).AddDays(-5);
-            Console.WriteLine(new { shipOn });
+
             order.Apply(new ShipOn(shipOn));
             await orderRepository.Save(order);
 
@@ -576,7 +576,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task A_command_handler_can_control_retries_of_a_failed_command()
+        public override async Task A_command_handler_can_control_retries_of_a_failed_command()
         {
             // arrange
             var order = CommandSchedulingTests_EventSourced.CreateOrder();
@@ -611,7 +611,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task A_command_handler_can_retry_a_failed_command_as_soon_as_it_wants()
+        public override async Task A_command_handler_can_request_retry_of_a_failed_command_as_soon_as_possible()
         {
             // arrange
             var order = CommandSchedulingTests_EventSourced.CreateOrder();
@@ -647,7 +647,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task A_command_handler_can_retry_a_failed_command_as_late_as_it_wants()
+        public override async Task A_command_handler_can_request_retry_of_a_failed_command_as_late_as_it_wants()
         {
             var order = CommandSchedulingTests_EventSourced.CreateOrder();
             order.Apply(
@@ -681,7 +681,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task The_aggregate_can_cancel_a_scheduled_command_after_it_fails()
+        public override async Task A_command_handler_can_cancel_a_scheduled_command_after_it_fails()
         {
             var order = new Order(
                 new CreateOrder(Any.FullName()))
@@ -759,7 +759,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task Specific_scheduled_commands_can_be_triggered_directly_by_aggregate_id()
+        public override async Task Specific_scheduled_commands_can_be_triggered_directly_by_target_id()
         {
             // arrange
             var shipmentId = Any.AlphanumericString(8, 8);
@@ -792,7 +792,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_command_is_triggering_and_succeeds_then_result_SuccessfulCommands_references_it()
+        public async Task When_a_command_is_triggered_and_succeeds_then_result_SuccessfulCommands_references_it()
         {
             // arrange
             var shipmentId = Any.AlphanumericString(8, 8);
@@ -820,7 +820,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_clock_is_advanced_then_result_SuccessfulCommands_are_included_in_the_result()
+        public async Task When_a_clock_is_advanced_then_resulting_SuccessfulCommands_are_included_in_the_result()
         {
             // arrange
             var shipmentId = Any.AlphanumericString(8, 8);
@@ -848,7 +848,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_triggering_specific_commands_then_the_result_can_be_used_to_evaluate_failures()
+        public override async Task When_triggering_specific_commands_then_the_result_can_be_used_to_evaluate_failures()
         {
             // arrange
             var shipmentId = Any.AlphanumericString(8, 8);
@@ -923,7 +923,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_command_is_scheduled_but_an_exception_is_thrown_in_a_handler_then_an_error_is_recorded()
+        public override async Task When_a_command_is_scheduled_but_an_exception_is_thrown_in_a_handler_then_an_error_is_recorded()
         {
             Configuration.Current.UseDependency(_ => new CustomerAccount.OrderEmailConfirmer
             {
@@ -991,7 +991,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_command_is_scheduled_but_the_aggregate_it_applies_to_is_not_found_then_the_command_is_retried()
+        public override async Task When_a_command_is_scheduled_but_the_target_it_applies_to_is_not_found_then_the_command_is_retried()
         {
             // create and cancel an order for a nonexistent customer account 
             var customerId = Any.Guid();
@@ -1105,7 +1105,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task Constructor_commands_can_be_scheduled_to_create_new_aggregate_instances()
+        public override async Task Constructor_commands_can_be_scheduled_to_create_new_aggregate_instances()
         {
             var commandScheduler = Configuration.Current.CommandScheduler<Order>();
 
@@ -1124,7 +1124,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_constructor_command_fails_with_a_ConcurrencyException_it_is_not_retried()
+        public override async Task When_a_constructor_command_fails_with_a_ConcurrencyException_it_is_not_retried()
         {
             var commandScheduler = Configuration.Current.CommandScheduler<Order>();
 
@@ -1157,7 +1157,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_an_immediately_scheduled_command_depends_on_an_event_that_has_not_been_saved_yet_then_there_is_not_initially_a_concurrency_exception()
+        public override async Task When_an_immediately_scheduled_command_depends_on_a_precondition_that_has_not_been_met_yet_then_there_is_not_initially_a_concurrency_exception()
         {
             var commandScheduler = Configuration.Current.CommandScheduler<Order>();
 
@@ -1240,7 +1240,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_a_scheduled_command_depends_on_an_event_that_never_arrives_it_is_eventually_abandoned()
+        public override async Task When_a_scheduled_command_depends_on_an_event_that_never_arrives_it_is_eventually_abandoned()
         {
             VirtualClock.Start();
             var commandScheduler = Configuration.Current.CommandScheduler<Order>();
@@ -1291,7 +1291,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         }
 
         [Test]
-        public async Task When_command_is_durable_but_immediate_delivery_succeeds_then_it_is_not_redelivered()
+        public override async Task When_command_is_durable_but_immediate_delivery_succeeds_then_it_is_not_redelivered()
         {
             var scheduleCount = 0;
             var deliverCount = 0;
